requirements.txt
python-telegram-bot[job-queue]
httpx
tzdata
